extract embeded png from svg files
